refactored code to make more generic
