+++ DNAseq/fastq_to_cram/requirements.txt
@@ -4,3 +4,4 @@ mypy-extensions==1.0.0
 packaging==24.2
 pathspec==0.12.1
 platformdirs==4.3.6
+setuptools==75.3.0
